@Query를 활용한 group by 쿼리로 변경

+++ src/test/kotlin/com/group/libraryapp/service/book/BookServiceTest.kt
@@ -129,16 +129,16 @@ class BookServiceTest @Autowired constructor (
 
         // then
         assertThat(results).hasSize(2)
-        assertCount(results, BookType.COMPUTER, 2)
+        assertCount(results, BookType.COMPUTER, 2L)
 //        val computerDto = results.first { result -> result.type == BookType.COMPUTER }
 //        assertThat(computerDto.count).isEqualTo(2)
 
-        assertCount(results, BookType.SCIENCE, 1)
+        assertCount(results, BookType.SCIENCE, 1L)
 //        val scienceDto = results.first { result -> result.type == BookType.SCIENCE }
 //        assertThat(scienceDto.count).isEqualTo(1)
     }
 
-    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Int) {
+    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Long) {
         assertThat(results.first { result -> result.type == type }.count).isEqualTo(count)
     }
 }
